+++ src/main/java/com/authlete/jaxrs/spi/AuthorizationDecisionHandlerSpiAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016-2019 Authlete, Inc.
+ * Copyright (C) 2016-2020 Authlete, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 package com.authlete.jaxrs.spi;
 
 
+import java.util.List;
 import com.authlete.common.assurance.VerifiedClaims;
 import com.authlete.common.assurance.constraint.VerifiedClaimsConstraint;
 import com.authlete.common.dto.Property;
@@ -86,7 +87,7 @@ public class AuthorizationDecisionHandlerSpiAdapter implements AuthorizationDeci
 
 
     @Override
-    public VerifiedClaims getVerifiedClaims(String subject, VerifiedClaimsConstraint constraint)
+    public List<VerifiedClaims> getVerifiedClaims(String subject, VerifiedClaimsConstraint constraint)
     {
         return null;
     }